Changelog - Fix error imports due to conflicted rebase
