+++ features/article/repository.go
@@ -8,7 +8,7 @@ import (
 
 type ArticleRepository interface {
 	Create(ctx context.Context, a *models.Article) (int64, error)
-	GetAll(ctx context.Context, cursor string, num int64) ([]*models.Article, error)
+	GetAll(ctx context.Context, cursor string, num int64) (*[]*models.Article, error)
 	GetByID(ctx context.Context, id int64) (*models.Article, error)
 	GetByTitle(ctx context.Context, title string) (*models.Article, error)
 	Update(ctx context.Context, article *models.Article) (*models.Article, error)
