+++ features/rating/delivery/httpecho/rating_handler.go
